Preferences: Reorganize UI and Theme pages in Display group
This changes current "Display/UI" pages to "Display/Advenced" as most of
settings there were pretty technical. These settings are not removed as
it is hard to decide which one should be removed and which one should
stay and we are too close to UI freeze.

Settings that should be obviously kept were moved to new "Display/UI"
section which now includes also options from previous "Display/Theme".

Fixes: #14412
