--- src/Gui/PreferencePages/DlgSettingsTheme.ui
@@ -1,224 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>Gui::Dialog::DlgSettingsTheme</class>
- <widget class="QWidget" name="Gui::Dialog::DlgSettingsTheme">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>405</width>
-    <height>400</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Theme</string>
-  </property>
-  <layout class="QVBoxLayout" name="verticalLayout">
-   <item>
-    <widget class="QLabel" name="label">
-     <property name="text">
-      <string>This page let you customize your current theme. The offered settings are optional for theme developers so they may or may not have an effect in your current theme.</string>
-     </property>
-     <property name="wordWrap">
-      <bool>true</bool>
-     </property>
-    </widget>
-   </item>
-   <item>
-    <layout class="QGridLayout">
-     <item row="0" column="0">
-      <layout class="QGridLayout">
-       <item row="0" column="0">
-        <widget class="QLabel" name="label1">
-         <property name="text">
-          <string>Accent color 1</string>
-         </property>
-        </widget>
-       </item>
-       <item row="0" column="1">
-        <widget class="Gui::PrefColorButton" name="ThemeAccentColor1">
-         <property name="sizePolicy">
-          <sizepolicy hsizetype="MinimumExpanding" vsizetype="Fixed">
-           <horstretch>0</horstretch>
-           <verstretch>0</verstretch>
-          </sizepolicy>
-         </property>
-         <property name="toolTip">
-          <string>This color might be used by your theme to let you customize it.</string>
-         </property>
-         <property name="color" stdset="0">
-          <color>
-           <red>85</red>
-           <green>123</green>
-           <blue>182</blue>
-          </color>
-         </property>
-         <property name="prefEntry" stdset="0">
-          <cstring>ThemeAccentColor1</cstring>
-         </property>
-         <property name="prefPath" stdset="0">
-          <cstring>Themes</cstring>
-         </property>
-        </widget>
-       </item>
-       <item row="1" column="0">
-        <widget class="QLabel" name="label2">
-         <property name="text">
-          <string>Accent color 2</string>
-         </property>
-        </widget>
-       </item>
-       <item row="1" column="1">
-        <widget class="Gui::PrefColorButton" name="ThemeAccentColor2">
-         <property name="sizePolicy">
-          <sizepolicy hsizetype="MinimumExpanding" vsizetype="Fixed">
-           <horstretch>0</horstretch>
-           <verstretch>0</verstretch>
-          </sizepolicy>
-         </property>
-         <property name="toolTip">
-          <string>This color might be used by your theme to let you customize it.</string>
-         </property>
-         <property name="color" stdset="0">
-          <color>
-           <red>85</red>
-           <green>123</green>
-           <blue>182</blue>
-          </color>
-         </property>
-         <property name="prefEntry" stdset="0">
-          <cstring>ThemeAccentColor2</cstring>
-         </property>
-         <property name="prefPath" stdset="0">
-          <cstring>Themes</cstring>
-         </property>
-        </widget>
-       </item>
-       <item row="2" column="0">
-        <widget class="QLabel" name="label3">
-         <property name="text">
-          <string>Accent color 3</string>
-         </property>
-        </widget>
-       </item>
-       <item row="2" column="1">
-        <widget class="Gui::PrefColorButton" name="ThemeAccentColor3">
-         <property name="sizePolicy">
-          <sizepolicy hsizetype="MinimumExpanding" vsizetype="Fixed">
-           <horstretch>0</horstretch>
-           <verstretch>0</verstretch>
-          </sizepolicy>
-         </property>
-         <property name="toolTip">
-          <string>This color might be used by your theme to let you customize it.</string>
-         </property>
-         <property name="color" stdset="0">
-          <color>
-           <red>85</red>
-           <green>123</green>
-           <blue>182</blue>
-          </color>
-         </property>
-         <property name="prefEntry" stdset="0">
-          <cstring>ThemeAccentColor3</cstring>
-         </property>
-         <property name="prefPath" stdset="0">
-          <cstring>Themes</cstring>
-         </property>
-        </widget>
-       </item>
-       <item row="3" column="0">
-        <widget class="QLabel" name="styleSheetLabel">
-         <property name="text">
-          <string>Style sheet (advanced):</string>
-         </property>
-        </widget>
-       </item>
-       <item row="3" column="1">
-        <widget class="Gui::PrefComboBox" name="StyleSheets">
-         <property name="toolTip">
-          <string>Style sheet how user interface will look like</string>
-         </property>
-         <property name="prefPath" stdset="0">
-          <cstring>MainWindow</cstring>
-         </property>
-         <property name="prefEntry" stdset="0">
-          <cstring>StyleSheet</cstring>
-         </property>
-         <property name="prefType" stdset="0">
-          <cstring></cstring>
-         </property>
-        </widget>
-       </item>
-       <item row="4" column="0">
-        <widget class="QLabel" name="label_2">
-         <property name="text">
-          <string>Overlay style sheet:</string>
-         </property>
-        </widget>
-       </item>
-       <item row="4" column="1">
-        <widget class="Gui::PrefComboBox" name="OverlayStyleSheets">
-         <property name="prefPath" stdset="0">
-          <cstring>MainWindow</cstring>
-         </property>
-         <property name="prefEntry" stdset="0">
-          <cstring>OverlayActiveStyleSheet</cstring>
-         </property>
-         <property name="prefType" stdset="0">
-          <cstring></cstring>
-         </property>
-        </widget>
-       </item>
-      </layout>
-     </item>
-     <item row="0" column="1">
-      <spacer name="spacer_5">
-       <property name="orientation">
-        <enum>Qt::Horizontal</enum>
-       </property>
-       <property name="sizeHint" stdset="0">
-        <size>
-         <width>183</width>
-         <height>23</height>
-        </size>
-       </property>
-      </spacer>
-     </item>
-    </layout>
-   </item>
-   <item>
-    <spacer name="spacer_3">
-     <property name="orientation">
-      <enum>Qt::Vertical</enum>
-     </property>
-     <property name="sizeHint" stdset="0">
-      <size>
-       <width>20</width>
-       <height>20</height>
-      </size>
-     </property>
-    </spacer>
-   </item>
-  </layout>
- </widget>
- <customwidgets>
-  <customwidget>
-   <class>Gui::ColorButton</class>
-   <extends>QPushButton</extends>
-   <header>Gui/Widgets.h</header>
-  </customwidget>
-  <customwidget>
-   <class>Gui::PrefColorButton</class>
-   <extends>Gui::ColorButton</extends>
-   <header>Gui/PrefWidgets.h</header>
-  </customwidget>
-  <customwidget>
-   <class>Gui::PrefComboBox</class>
-   <extends>QComboBox</extends>
-   <header>Gui/PrefWidgets.h</header>
-  </customwidget>
- </customwidgets>
- <resources/>
- <connections/>
-</ui>
